refactor(nf): move VectorTokenizer/ImageTokenizer to models/core
